Move create billing batch service
